pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
discord = "^1.7.3"
google-api-python-client = "^2.41.0"
python-dotenv = "^0.19.2"
requests = "^2.27.1"
youtube_dl = "^2021.12.17"
PyNaCl = "^1.5.0"
Pillow = "^9.0.1"
ffmpeg = "^1.4"
opus20 = "^1.0.1"
Flask = "^2.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"